Docs: Base UI first — /docs renders Base UI, /docs/radix the classic variant
Replace the two parallel docs implementations (Radix-flavored
docs/page.tsx + color-only variant-docs.tsx) with one variant-
parameterized FullDocs:

- Every live example resolves its picker namespaces (CP/GP/FP) from a
  variant context; code snippets stay authored against the Radix install
  paths and are path-swapped to fill-picker-base for the Base UI page.
- /docs = FullDocs base, /docs/radix = FullDocs radix, VariantToggle in
  the header; /docs/base now redirects to /docs.
- Install section leads with fill-picker-base.json on the Base UI page;
  the three Radix entry points remain on the Radix page.
- llms.txt / docs.md / Copy-for-AI markdown now recommend the Base UI
  variant first (REGISTRY_URL -> fill-picker-base.json).
- Delete superseded src/app/docs/variant-docs.tsx.
