+++ README.md
@@ -18,7 +18,7 @@ pnpm dlx shadcn@latest add https://amplo.ale.design/r/fill-picker-base.json
 
 The first three install into the same `components/ui/fill-picker/` tree, so picking a deeper entry point later is idempotent — shadcn dedupes by `target`.
 
-Two variants, one engine: the **Base UI** variant rebuilds the interactive parts (Hue/Lightness/Alpha sliders, FormatSwitcher, ChannelInput, Swatches, and the gradient TypeSwitcher/InterpSwitcher/StopList and friends) on [Base UI](https://base-ui.com) primitives and is the main variant going forward; the classic variant keeps the original Radix-backed shadcn parts. Both share the same OKLCH engine, compound API, and presentational parts, so fixes propagate to both. Base UI demo: [/docs/base](https://amplo.ale.design/docs/base).
+Two variants, one engine: the **Base UI** variant rebuilds the interactive parts (Hue/Lightness/Alpha sliders, FormatSwitcher, ChannelInput, Swatches, and the gradient TypeSwitcher/InterpSwitcher/StopList and friends) on [Base UI](https://base-ui.com) primitives and is the main variant going forward; the classic variant keeps the original Radix-backed shadcn parts. Both share the same OKLCH engine, compound API, and presentational parts, so fixes propagate to both. The docs default to Base UI at [`/docs`](https://amplo.ale.design/docs); the Radix variant lives at [`/docs/radix`](https://amplo.ale.design/docs/radix), one click away via the toggle.
 
 > Replace the URL with your deployed origin if you fork this repo. The registry artifacts are generated by `pnpm registry:build` and served from `public/r/`.
